Redid cors, setup page routing
