+++ backend/server.js
@@ -4,11 +4,8 @@ const mongoose = require("mongoose")
 const express = require("express")
 const app = express()
 const cors = require("cors")
-const corsOptions = {
-    origin: ["http://localhost:5173"]
-}
 
-app.use(cors(corsOptions))
+app.use(cors({origin: '*'}))
 
 app.get("/api", (req, res) => {
     res.json({"dummy": ["testing", "the", "backend"]})
