+++ frontend/src/App.tsx
@@ -1,45 +1,20 @@
-import { useState, useEffect } from 'react'
+import React from 'react'
 import reactLogo from './assets/react.svg'
 import viteLogo from '/vite.svg'
 import './App.css'
 import axios from "axios"
 import {Routes, Route} from 'react-router-dom'
 
+import Landing from './Components/Landing'
+
 function App() {
-  const [count, setCount] = useState(0)
-
-  const fetchAPI = async () => {
-    const res = await axios.get("http://localhost:8080/api")
-    console.log(res.data.dummy)
-  }
-
-  useEffect(() => {
-    fetchAPI()
-  },[])
 
   return (
-    <>
-      <div>
-        <a href="https://vitejs.dev" target="_blank">
-          <img src={viteLogo} className="logo" alt="Vite logo" />
-        </a>
-        <a href="https://react.dev" target="_blank">
-          <img src={reactLogo} className="logo react" alt="React logo" />
-        </a>
-      </div>
-      <h1>Vite + React</h1>
-      <div className="card">
-        <button onClick={() => setCount((count) => count + 1)}>
-          count is {count}
-        </button>
-        <p>
-          Edit <code>src/App.tsx</code> and save to test HMR
-        </p>
-      </div>
-      <p className="read-the-docs">
-        Click on the Vite and React logos to learn more
-      </p>
-    </>
+    <React.Fragment>
+      <Routes>
+        <Route path = '/' element={<Landing/>}/>
+      </Routes>
+    </React.Fragment>
   )
 }
